+++ app/containers/AgentDesktop/reducer.js
@@ -10,7 +10,9 @@
 
 import { fromJS, Map, List } from 'immutable';
 
-import Interaction, { activeContactFormBlank } from 'models/Interaction/Interaction';
+import Interaction, {
+  activeContactFormBlank,
+} from 'models/Interaction/Interaction';
 import Message from 'models/Message/Message';
 import ResponseMessage from 'models/Message/ResponseMessage';
 
@@ -487,6 +489,21 @@ function agentDesktopReducer(state = initialState, action) {
                 .set('onHold', action.response.customerOnHold === true)
                 .set('recording', action.response.recording === true);
             }
+            if (
+              action.newStatus === 'work-accepting' &&
+              updatedInteraction.get('isScriptOnly') === true
+            ) {
+              // Remove isScriptOnly if we are accepting the work offer
+              updatedInteraction = updatedInteraction.delete('isScriptOnly');
+              // Keep script in focus (voice will always be in focus, it's script is in MainContentArea)
+              if (updatedInteraction.get('channelType') !== 'voice') {
+                updatedInteraction = updatedInteraction.set(
+                  'sidePanelTabIndex',
+                  2
+                );
+                openContactsPanel = true;
+              }
+            }
             return updatedInteraction;
           })
           .set(
@@ -682,13 +699,17 @@ function agentDesktopReducer(state = initialState, action) {
             action.response.channelType === 'email')
         )
       ) {
-        // If interaction was already added by START_OUTBOUND_INTERACTION, replace it; otherwise, just push it to the list
+        // If interaction was already added by START_OUTBOUND_INTERACTION or ADD_SCRIPT, replace it; otherwise, just push it to the list
         const interactionIndex = state
           .get('interactions')
           .findIndex(
             (interaction) =>
-              interaction.get('direction') === 'outbound' &&
-              interaction.get('channelType') === action.response.channelType
+              (interaction.get('direction') === 'outbound' &&
+                interaction.get('channelType') ===
+                  action.response.channelType) ||
+              (interaction.get('interactionId') ===
+                action.response.interactionId &&
+                interaction.get('status') === 'script-only')
           );
         const interactionToAdd = new Map(new Interaction(action.response));
         if (interactionIndex !== -1) {
@@ -729,12 +750,47 @@ function agentDesktopReducer(state = initialState, action) {
             interaction.get('interactionId') === action.interactionId
         );
       if (interactionIndex > -1) {
-        return state.setIn(
-          ['interactions', interactionIndex, 'script'],
-          fromJS(action.script)
+        let openContactsPanel = false;
+        let newState = state.updateIn(
+          ['interactions', interactionIndex],
+          (interaction) => {
+            const newInteraction = interaction.set(
+              'script',
+              fromJS(action.script)
+            );
+            // Put script in focus (voice will always be in focus, it's script is in MainContentArea)
+            if (interaction.get('channelType') !== 'voice') {
+              openContactsPanel = true;
+              return newInteraction.set('sidePanelTabIndex', 2);
+            } else {
+              return newInteraction;
+            }
+          }
         );
+        if (openContactsPanel) {
+          newState = newState.set('isContactsPanelCollapsed', false);
+        }
+        return newState;
       } else {
-        return state;
+        // 'script-only' is the main status we will use. isScriptOnly for when interactions receive a work offer, but still need to render the script in MainContentArea until it has been accepted
+        const scriptInteraction = fromJS({
+          interactionId: action.interactionId,
+          status: 'script-only',
+          isScriptOnly: true,
+          script: action.script,
+          sidePanelTabIndex: 0,
+          query: {},
+          contact: {},
+          activeContactForm: activeContactFormBlank,
+        });
+        return state
+          .update('interactions', (interactions) =>
+            interactions.push(scriptInteraction)
+          )
+          .set(
+            'selectedInteractionId',
+            state.get('selectedInteractionId') || action.interactionId
+          );
       }
     }
     case REMOVE_SCRIPT: {
@@ -745,10 +801,15 @@ function agentDesktopReducer(state = initialState, action) {
             interaction.get('interactionId') === action.interactionId
         );
       if (interactionIndex > -1) {
-        // Remove the interaction if the script is the only thing left to do
+        const interactionStatus = state.getIn([
+          'interactions',
+          interactionIndex,
+          'status',
+        ]);
+        // Remove the interaction if the script is the only thing to do
         if (
-          state.getIn(['interactions', interactionIndex, 'status']) ===
-          'work-ended-pending-script'
+          interactionStatus === 'work-ended-pending-script' ||
+          interactionStatus === 'script-only'
         ) {
           return removeInteractionAndSetNextSelectedInteraction(
             state,
@@ -825,9 +886,10 @@ function agentDesktopReducer(state = initialState, action) {
             (interaction) =>
               interaction.get('interactionId') === action.interactionId
           );
-        return state.setIn(
-          ['interactions', interactionIndex, 'status'],
-          'work-ended-pending-script'
+        return state.updateIn(['interactions', interactionIndex], (interaction) =>
+          interaction
+            .set('status', 'work-ended-pending-script')
+            .set('contactMode', 'view')
         );
       } else {
         return state;
@@ -1322,11 +1384,12 @@ function agentDesktopReducer(state = initialState, action) {
             return interaction.update('warmTransfers', (warmTransfers) =>
               warmTransfers.map((warmTransfer) => {
                 if (warmTransfer.get('id') === action.response.result.id) {
-                  const name = action.response.result.firstName ||
+                  const name =
+                    action.response.result.firstName ||
                     action.response.result.lastName
-                    ? `${action.response.result.firstName} ${action.response
-                        .result.lastName}`
-                    : action.response.result.email;
+                      ? `${action.response.result.firstName} ${action.response
+                          .result.lastName}`
+                      : action.response.result.email;
                   return warmTransfer.set('name', name);
                 } else {
                   return warmTransfer;