+++ app/containers/Interaction/index.js
@@ -24,7 +24,6 @@ import Progress from 'components/Progress';
 import Button from 'components/Button';
 import { cancelClickToDial } from 'containers/AgentDesktop/actions';
 
-import { selectAwaitingDisposition } from 'containers/AgentDesktop/selectors';
 import { selectActiveExtension } from 'containers/AgentStatusMenu/selectors';
 
 import messages from './messages';
@@ -153,10 +152,7 @@ export class Interaction extends React.Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (
-      this.props.status !== nextProps.status ||
-      this.props.previewText !== nextProps.previewText
-    ) {
+    if (this.props.status !== 'wrapup' && nextProps.status === 'wrapup') {
       this.setState({
         startTime: Date.now(),
         ageSeconds: 0,
@@ -277,7 +273,42 @@ export class Interaction extends React.Component {
   };
 
   render() {
-    if (this.props.status !== 'creating-new-interaction') {
+    if (
+      this.props.status === 'creating-new-interaction' ||
+      this.props.status === 'script-only'
+    ) {
+      return (
+        <div
+          id={`${this.props.status}InteractionContainer-${this.props
+            .interactionId}`}
+          className={`${this.props.status}InteractionContainer`}
+          style={[
+            styles.base,
+            this.props.selected && styles.selectedBase,
+            styles.newInteractionBase,
+          ]}
+          key={this.props.interactionId}
+          onClick={this.props.onClick}
+          disabled={this.props.selected}
+        >
+          <div style={styles.iconContainer} />
+          <div
+            style={[
+              styles.mainContainer,
+              styles.mainContainer[this.props.status],
+            ]}
+          >
+            <div style={styles.headerContainer}>
+              <div style={styles.from}>
+                {this.props.status === 'creating-new-interaction'
+                  ? <FormattedMessage {...messages.newInteraction} />
+                  : <FormattedMessage {...messages.script} />}
+              </div>
+            </div>
+          </div>
+        </div>
+      );
+    } else {
       const pendingPSTN =
         this.props.activeExtension.type === 'pstn' &&
         this.props.status === 'pending' &&
@@ -321,7 +352,7 @@ export class Interaction extends React.Component {
               ? <div style={styles.intentText}>
                 <FormattedMessage {...acceptMessage} />
                 {this.props.interactionDirection === 'outbound' &&
-                    this.props.channelType === 'voice'
+                  this.props.channelType === 'voice'
                     ? <Button
                       id="cancelInteractionBeforeActive"
                       type="primaryRed"
@@ -332,36 +363,6 @@ export class Interaction extends React.Component {
                     : undefined}
               </div>
               : undefined}
-          </div>
-        </div>
-      );
-    } else {
-      return (
-        <div
-          id={`${this.props.status}InteractionContainer-${this.props
-            .interactionId}`}
-          className={`${this.props.status}InteractionContainer`}
-          style={[
-            styles.base,
-            this.props.selected && styles.selectedBase,
-            styles.newInteractionBase,
-          ]}
-          key={this.props.interactionId}
-          onClick={this.props.onClick}
-          disabled={this.props.selected}
-        >
-          <div style={styles.iconContainer} />
-          <div
-            style={[
-              styles.mainContainer,
-              styles.mainContainer[this.props.status],
-            ]}
-          >
-            <div style={styles.headerContainer}>
-              <div style={styles.from}>
-                New Interaction
-              </div>
-            </div>
           </div>
         </div>
       );
@@ -384,8 +385,9 @@ Interaction.propTypes = {
     'wrapup',
     'creating-new-interaction',
     'work-ended-pending-script',
+    'script-only',
   ]).isRequired,
-  awaitingDisposition: PropTypes.bool.isRequired,
+  awaitingDisposition: PropTypes.bool,
   selected: PropTypes.bool,
   onClick: PropTypes.func,
   activeExtension: PropTypes.object.isRequired,
@@ -395,7 +397,6 @@ Interaction.propTypes = {
 };
 
 const mapStateToProps = (state, props) => ({
-  awaitingDisposition: selectAwaitingDisposition(state, props),
   activeExtension: selectActiveExtension(state, props),
 });